add cli, fix presenter
