+++ src/main/kotlin/presentation/presenter/project/DeleteProjectPresenter.kt
@@ -8,27 +8,7 @@ import java.util.UUID
 class DeleteProjectPresenter(
     private val deleteProjectUseCase: DeleteProjectUseCase
 ) {
-    operator fun invoke(){
-        println("DELETE PROJECT")
-        val input = readln().trim()
-        val projectId = UUID.fromString(input)
-        deleteProjectUseCase.deleteProject(projectId).fold(
-            onSuccess = { project->
-                showDeletedProjectDetails(project)
-            },
-            onFailure = { throwable ->
-                println(throwable.message)
-            }
-        )
-
-    }
-    private fun showDeletedProjectDetails(project: Project){
-        println("the project id is ${project.projectId}" )
-        println("the project name is ${project.projectName}")
-        println("the project description is ${ project.projectDescription }")
-        println("the responsible admin Id is: ${project.adminId}")
-        println("the project was created at: ${project.createdAt}")
-        println("the task state is: ${project.taskStates}")
-
+    fun deleteProject(projectId : UUID): Result<Project>{
+        return deleteProjectUseCase.deleteProject(projectId)
     }
 }
