+++ src/main/kotlin/presentation/view/project/DeleteProjectCLI.kt
@@ -1,0 +1,31 @@
+package org.example.presentation.view.project
+
+import org.example.domain.model.Project
+import org.example.presentation.presenter.project.DeleteProjectPresenter
+import java.util.UUID
+
+class DeleteProjectCLI (private val deleteProjectPresenter: DeleteProjectPresenter){
+    operator fun invoke(projectId1: UUID) {
+        println("DELETE PROJECT")
+        val input = readln().trim()
+        val projectId = UUID.fromString(input)
+        deleteProjectPresenter.deleteProject(projectId).fold(
+            onSuccess = { project->
+                showDeletedProjectDetails(project)
+            },
+            onFailure = { throwable ->
+                println(throwable.message)
+            }
+        )
+
+    }
+    private fun showDeletedProjectDetails(project: Project){
+        println("the project id is ${project.projectId}" )
+        println("the project name is ${project.projectName}")
+        println("the project description is ${ project.projectDescription }")
+        println("the responsible admin Id is: ${project.adminId}")
+        println("the project was created at: ${project.createdAt}")
+        println("the task state is: ${project.taskStates}")
+
+    }
+}
